MP: Use fs_searchpaths to load VM DLLs. Restore code for loading QVMs.
Native is always preferred over QVM when in same pk3 or if not a pure server and native/qvm are unzipped. However, if vm_* is more than 0 and QVM is in a pk3 after native, QVM will be used. Currently natives and qvms need to be in same pk3 for sv_pure to work when using qvms.

vm_game, vm_cgame, and vm_ui cvars slightly changed from Quake 3.
0 - load native only
1 - load native or interpreted QVM
2 - load native or compiled QVM