MP: Don't load unzipped dlls from homepath
Unzipped dlls in homepath are extracted from pk3s. Ignore them. If dlls are found in a pk3, the ones on disk in homepath will be checked/overwritten and then used. This way unpure VMs can be choosen based on pk3s, which may have QVMs that should override DLLs if vm_xxx 2.