All: Don't warn when weapon cfg is missing altswitch animations
The parser was recently fixed to detect missing tokens. This caused
an error message for some of the default MP weapon cfgs that do not
include altswitch animations. None of the weapons actually use the
altswitch animation so disable the warning for them.

Previously the missing animations would be loaded with 0 for all
params. I just copy idle1 to them instead. Shouldn't be used in
any case.