All: Fix loading map from command line in SP
map command unloads the UI VM and then cinematic command is processed.
Video is not displayed because UI VM is not loaded and input causes
error that UI VM is NULL. I changed it to use Quake 3 behavior of only
playing intro if there are no command line arguments. This does not
affect MP as there is no logo or intro video present.

Fixes running "iowolfsp.x86_64 +spmap escape1".