All: Fix manual pick up of chalice at end of crypt2
The chalice at the end of crypt2 single-player map is inside of
a trigger_hurt entity. If cg_autoactivate is 1 (the default), the
chalice is picked up when touched. However with cg_autoactivate 0
it was not possible to pick up (activate) the chalice because game
tries to activate the trigger_hurt entity.

Make hint cursor and +activate trace skip trigger_hurt entity.