All: Rend2: Fix specular on barrels in mp_base
The RTCW MP shader models\furniture\barrel\barrel_m01 has a lightmap
stage and it's used on a game model models/furniture/barrel/barrel_d.mdc.

"map $lightmap" on game models was incorrectly applying specular due
to incorrect vertex lightdir (which is not the vertex normal).

This commit ignores the white lightmaps on models for the lightall GLSL
shader. This looks fine for the barrel in mp_base.

This change prevents r_forceSun 1 from drawing sun shadows on game
models with a lightmap. Normal map can't be used but it probably doesn't
work anyway on game models with a lightmap as they typically have vertex
lightdir of 0,0,0. So this fixes it but maybe not in the best way.

I think incorrect vertex lightdir may happen when mixing CPU animated
BSP surfaces and models (due to not entirely setting the tess geometry)
in ioquake3. This is more of a problem in iortcw due to offsetting the
vertex lightdir based on the normal in the GLSL so it's never 0,0,0.